% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_pipe.R
\name{insert_pipe_magrittr}
\alias{insert_pipe_magrittr}
\title{Insert Pipe Operation.}
\usage{
insert_pipe_magrittr()
}
\description{
Call this function as an addin to insert \code{ \%>\% } at the cursor position.
}
